// src/src/enums.ts
export enum gameTypes {
  SinglePlayer = "Single-player",
  Multiplayer = "Multi-player",
}

export enum gameGenre {
  RPG = "Role-Playing Game (RPG)",
  RythmGame = "Rythm Game",
  ActionAdventure = "Action-Adventure",
  Platformer = "Platformer",
  Puzzle = "Puzzle",
  Simulation = "Simulation",
  Strategy = "Strategy",
  Horror = "Horror",
  VisualNovel = "Visual Novel",
  Metroidvania = "Metroidvania",
  Sandbox = "Sandbox",

  FPS = "First-Person Shooter (FPS)",
  BattleRoyale = "Battle Royale",
  MOBA = "MOBA",
  MMORPG = "MMORPG",
  Fighting = "Fighting",
  Sports = "Sports",
  Racing = "Racing",
  PartyGame = "Party Game",
  CoOpSurvival = "Co-op Survival",
  RTS = "Real-Time Strategy (RTS)",
}

export enum coverTitlePosition {
  Top = "top",
  Center = "center",
  Bottom = "bottom",
}

export enum pageType{
  Home = "home",
  About = "about",
  Game = "game"
}